requirements.txt
Flask==2.3.2
flask-cors==3.0.10
pyodbc==4.0.35
pandas==2.0.3
python-dotenv==1.0.1